+++ sdk/ai/azure-ai-agents/azure/ai/agents/models/_patch.py
@@ -67,6 +67,14 @@ from ._models import (
     RunStepDeltaChunk,
     BingGroundingSearchConfiguration,
     BingGroundingSearchToolParameters,
+    BingCustomSearchToolDefinition,
+    BingCustomSearchConfiguration,
+    BingCustomSearchToolParameters,
+    SharepointToolDefinition,
+    SharepointGroundingToolParameters,
+    ToolConnection,
+    MicrosoftFabricToolDefinition,
+    FabricDataAgentToolParameters,
     SubmitToolOutputsAction,
     ThreadRun,
     ToolDefinition,
@@ -841,6 +849,33 @@ class AzureFunctionTool(Tool[AzureFunctionToolDefinition]):
         pass
 
 
+class ConnectionTool(Tool[ToolDefinitionT]):
+    """
+    A tool that requires connection ids.
+    Used as base class for Bing Grounding, Sharepoint, and Microsoft Fabric
+    """
+
+    def __init__(self, connection_id: str):
+        """
+        Initialize ConnectionTool with a connection_id.
+
+        :param connection_id: Connection ID used by tool. All connection tools allow only one connection.
+        """
+        self.connection_ids = [ToolConnection(connection_id=connection_id)]
+
+    @property
+    def resources(self) -> ToolResources:
+        """
+        Get the connection tool resources.
+
+        :rtype: ToolResources
+        """
+        return ToolResources()
+
+    def execute(self, tool_call: Any) -> Any:
+        pass
+
+
 class BingGroundingTool(Tool[BingGroundingToolDefinition]):
     """
     A tool that searches for information using Bing.
@@ -848,13 +883,14 @@ class BingGroundingTool(Tool[BingGroundingToolDefinition]):
 
     def __init__(self, connection_id: str, market: str = "", set_lang: str = "", count: int = 5, freshness: str = ""):
         """
-        Initialize Bing Custom Search with a connection_id.
+        Initialize Bing Grounding tool with a connection_id.
 
-        :param connection_id: Connection ID used by tool. Bing Custom Search tools allow only one connection.
-        :param market:
-        :param set_lang:
-        :param count:
-        :param freshness:
+        :param connection_id: Connection ID used by tool. Bing Grounding tools allow only one connection.
+        :param market: The market where the results come from.
+        :param set_lang: The language to use for user interface strings when calling Bing API.
+        :param count: The number of search results to return in the Bing API response.
+        :param freshness: Filter search results by a specific time range.
+
         """
         self.connection_ids = [
             BingGroundingSearchConfiguration(
@@ -886,6 +922,105 @@ class BingGroundingTool(Tool[BingGroundingToolDefinition]):
 
     def execute(self, tool_call: Any) -> Any:
         pass
+
+
+class BingCustomSearchTool(Tool[BingCustomSearchToolDefinition]):
+    """
+    A tool that searches for information using Bing Custom Search.
+    """
+
+    def __init__(
+        self,
+        connection_id: str,
+        instance_name: str,
+        market: str = "",
+        set_lang: str = "",
+        count: int = 5,
+        freshness: str = "",
+    ):
+        """
+        Initialize Bing Custom Search with a connection_id.
+
+        :param connection_id: Connection ID used by tool. Bing Custom Search tools allow only one connection.
+        :param instance_name: Config instance name used by tool.
+        :param market: The market where the results come from.
+        :param set_lang: The language to use for user interface strings when calling Bing API.
+        :param count: The number of search results to return in the Bing API response.
+        :param freshness: Filter search results by a specific time range.
+        """
+        self._search_configurations = [
+            BingCustomSearchConfiguration(
+                connection_id=connection_id,
+                instance_name=instance_name,
+                market=market,
+                set_lang=set_lang,
+                count=count,
+                freshness=freshness,
+            )
+        ]
+
+    @property
+    def definitions(self) -> List[BingCustomSearchToolDefinition]:
+        """
+        Get the Bing Custom Search tool definitions.
+
+        :rtype: List[ToolDefinition]
+        """
+        return [
+            BingCustomSearchToolDefinition(
+                bing_custom_search=BingCustomSearchToolParameters(search_configurations=self._search_configurations)
+            )
+        ]
+
+    @property
+    def resources(self) -> ToolResources:
+        """
+        Get the tool resources.
+
+        :rtype: ToolResources
+        """
+        return ToolResources()
+
+    def execute(self, tool_call: Any) -> Any:
+        pass
+
+
+class FabricTool(ConnectionTool[MicrosoftFabricToolDefinition]):
+    """
+    A tool that searches for information using Microsoft Fabric.
+    """
+
+    @property
+    def definitions(self) -> List[MicrosoftFabricToolDefinition]:
+        """
+        Get the Microsoft Fabric tool definitions.
+
+        :rtype: List[ToolDefinition]
+        """
+        return [
+            MicrosoftFabricToolDefinition(
+                fabric_dataagent=FabricDataAgentToolParameters(connection_list=self.connection_ids)
+            )
+        ]
+
+
+class SharepointTool(ConnectionTool[SharepointToolDefinition]):
+    """
+    A tool that searches for information using Sharepoint.
+    """
+
+    @property
+    def definitions(self) -> List[SharepointToolDefinition]:
+        """
+        Get the Sharepoint tool definitions.
+
+        :rtype: List[ToolDefinition]
+        """
+        return [
+            SharepointToolDefinition(
+                sharepoint_grounding=SharepointGroundingToolParameters(connection_list=self.connection_ids)
+            )
+        ]
 
 
 class ConnectedAgentTool(Tool[ConnectedAgentToolDefinition]):
@@ -1703,6 +1838,9 @@ __all__: List[str] = [
     "FunctionTool",
     "OpenApiTool",
     "BingGroundingTool",
+    "FabricTool",
+    "SharepointTool",
+    "BingCustomSearchTool",
     "StreamEventData",
     "AzureAISearchTool",
     "Tool",