+++ sdk/ai/azure-ai-agents/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugs Fixed
 
+* Adding instrumentation for create_thread_and_run
+
 ## 1.0.0 (2025-05-15)
 
 ### Features Added